requirements.txt
flask==3.0.2
flask-cors==4.0.0
yagmail==0.15.293
python-dotenv==1.0.1